requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djangochannelsrestframework
django-redis
hiredis
psycopg2-binary
channels
channels-redis
drf_recaptcha
tencentcloud-sdk-python
six
https://github.com/chibisov/drf-extensions/archive/master.zip
docker
django-crontab